Restructure indexscan API (index_beginscan, index_getnext) per yesterday's proposal to pghackers.  Also remove unnecessary parameters to heap_beginscan, heap_rescan.  I modified pg_proc.h to reflect the new numbers of parameters for the AM interface routines, but did not force an initdb because nothing actually looks at those fields.

+++ src/include/access/rtree.h
@@ -71,7 +71,7 @@ typedef struct RTreeScanOpaqueData
 	struct RTSTACK *s_stack;
 	struct RTSTACK *s_markstk;
 	uint16		s_flags;
-	uint16		s_internalNKey;
+	int			s_internalNKey;
 	ScanKey		s_internalKey;
 } RTreeScanOpaqueData;
